Package/deploy to pypi
- Fix setup.py to build package
- Split in command line executable and main logic
- Use `appdirs` to get user config dir
- Update Readme
- Add travis build template
