+++ .travis.yml
@@ -1,0 +1,17 @@
+language: python
+python:
+    - "2.7"
+
+install:
+    - pip install -r requirements.txt
+
+script:
+    - pytest
+
+deploy:
+    provider: pypi
+    user: "your user"
+    password: "your password (replace with `travis encrypt 'your password' --add deploy.password`)"
+    on:
+        - branch: master
+        - tags: true
