+++ SSHKeyDistribut0r/__init__.py
@@ -1,0 +1,1 @@
+from .key_distribut0r import main
