+++ requirements.txt
@@ -1,5 +1,6 @@
+appdirs==1.4.0
 ecdsa==0.13
 paramiko==1.16.0
 pycrypto==2.6.1
-scp==0.10.2
 pyyaml==3.11
+scp==0.10.2
